linux i386 fix

+++ sigar_linux.go
@@ -91,9 +91,9 @@ func (self *Swap) Get() error {
 		return err
 	}
 
-	self.Total = sysinfo.Totalswap
-	self.Free = sysinfo.Freeswap
-	self.Used = self.Total - self.Free
+	self.Total = uint64(sysinfo.Totalswap)
+	self.Free  = uint64(sysinfo.Freeswap)
+	self.Used  = self.Total - self.Free
 
 	return nil
 }
